+++ src/main/kotlin/core/model/EmergencyDTO.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.SerialName
 
 @Serializable
 data class EmergencyDTO(
-    @SerialName("t") val endTime: Int = 0,
+    @SerialName("t") val endTime: Long = 0,
     @SerialName("r") val rewardOrCosts: List<ConfigOutputDTO>? = null,
     @SerialName("s") val status: String = "",
     @SerialName("c") val config: ConfigEmergencyDTO = ConfigEmergencyDTO()
