+++ src/main/kotlin/core/model/EmergencyRequirementDTO.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class EmergencyRequirementDTO(
-    @SerialName("pr") val configProfSpecializationId: Int = 0,
+    @SerialName("pr") val configProfSpecializationId: Long = 0,
     @SerialName("nr") val slotNumber: Int = 0,
 )
