+++ src/main/kotlin/core/model/PhaseDTO.kt
@@ -5,9 +5,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class PhaseDTO(
-    @SerialName("pid") val phaseId: Int = 0,
-    @SerialName("et") val expirationTime: Int = 0,
-    @SerialName("tl") val timeLeft: Int = 0,
-    @SerialName("it") val initTime: Int = 0,
-    @SerialName("myst") val mystId: Int = 0,
+    @SerialName("pid") val phaseId: Long = 0,
+    @SerialName("et") val expirationTime: Long = 0,
+    @SerialName("tl") val timeLeft: Long = 0,
+    @SerialName("it") val initTime: Long = 0,
+    @SerialName("myst") val mystId: Long = 0,
 )
