+++ src/main/kotlin/core/model/ResourceDTO.kt
@@ -21,9 +21,9 @@ import core.model.config.ConfigDTO
  */
 @Serializable
 data class ResourceDTO(
-    @SerialName("id") val id: Int = 0,
-    @SerialName("c") val configId: Int = 0,
-    @SerialName("a") val amount: Int = 0,
-    @SerialName("ac") val amountCap: Int = 0,
+    @SerialName("id") val id: Long = 0,
+    @SerialName("c") val configId: Long = 0,
+    @SerialName("a") val amount: Long = 0,
+    @SerialName("ac") val amountCap: Long = 0,
     @SerialName("imagePath") val imagePath: String = "",
 )
