+++ src/main/kotlin/core/model/config/ConfigCinemaDTO.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ConfigCinemaDTO(
     @SerialName("mv") val maxViewsPerDay: UInt = 0u,
-    @SerialName("mt") val maxResetLifeTime: Int = 0,
+    @SerialName("mt") val maxResetLifeTime: Long = 0,
 )
