+++ src/main/kotlin/core/model/config/ConfigCityWheelBuffDTO.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigCityWheelBuffDTO(
-    @SerialName("i") val id: Int = 0,
-    @SerialName("g") val gfxId: Int = 0,
+    @SerialName("i") val id: Long = 0,
+    @SerialName("g") val gfxId: Long = 0,
     @SerialName("l") val localeId: String = "",
     @SerialName("o") val specialOutput: ConfigOutputDTO? = null,
 )
