+++ src/main/kotlin/core/model/config/ConfigEmergencyDTO.kt
@@ -7,9 +7,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigEmergencyDTO(
-    @SerialName("l") val lifetime: Int = 0,
-    @SerialName("gid") val gfxId: Int = 0,
-    @SerialName("lid") val locaId: Int = 0,
+    @SerialName("l") val lifetime: Long = 0,
+    @SerialName("gid") val gfxId: Long = 0,
+    @SerialName("lid") val locaId: Long = 0,
     @SerialName("d") val departmentRequirements: List<DepartmentRequirementDTO> = emptyList(),
     @SerialName("er") val emergencyRequirements: List<EmergencyRequirementDTO> = emptyList(),
 )
