+++ src/main/kotlin/core/model/config/ConfigGoodDTO.kt
@@ -5,10 +5,10 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigGoodDTO(
-    @SerialName("t") val tagIds: List<Int> = emptyList(),
-    @SerialName("eqi") val associatedEventIds: List<Int> = emptyList(),
-    @SerialName("id") val configId: Int = 0,
-    @SerialName("l") val levelId: Int = 0,
+    @SerialName("t") val tagIds: List<Long> = emptyList(),
+    @SerialName("eqi") val associatedEventIds: List<Long> = emptyList(),
+    @SerialName("id") val configId: Long = 0,
+    @SerialName("l") val levelId: Long = 0,
     @SerialName("s") val sellable: Boolean = false,
     @SerialName("p") val produceable: Boolean = false,
     @SerialName("g") val giftable: Boolean = false,
