+++ src/main/kotlin/core/model/config/ConfigGroundRestrictionShapePointDTO.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigGroundRestrictionShapePointDTO(
-    @SerialName("id") val id: Int = 0,
-    @SerialName("o") val shapePosition: Int = 0,
-    @SerialName("x") val x: Int = 0,
-    @SerialName("y") val y: Int = 0,
+    @SerialName("id") val id: Long = 0,
+    @SerialName("o") val shapePosition: Long = 0,
+    @SerialName("x") val x: Long = 0,
+    @SerialName("y") val y: Long = 0,
 )
