+++ src/main/kotlin/core/model/config/ConfigImprovementDTO.kt
@@ -5,12 +5,12 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigImprovementDTO(
-    @SerialName("cid") val id: Int = 0,
+    @SerialName("cid") val id: Long = 0,
     @SerialName("gfx") val gfxId: Int = 0,
     @SerialName("lo") val locaId: String = "",
     @SerialName("q") val quality: String = "",
-    @SerialName("t") val lifetime: Int? = null,
+    @SerialName("t") val lifetime: Long? = null,
     @SerialName("c") val renewCost: Int = 0,
     @SerialName("eff") val effects: List<ConfigEffectDTO> = emptyList(),
-    @SerialName("tag") val tagIds: List<Int> = emptyList(),
+    @SerialName("tag") val tagIds: List<Long> = emptyList(),
 )
