+++ src/main/kotlin/core/model/config/ConfigInventoryBagDTO.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigInventoryBagDTO(
-    @SerialName("id") val id: Int = 0,
+    @SerialName("id") val id: Long = 0,
     @SerialName("is") val initialSlots: Int = 0,
     @SerialName("ms") val maxSlots: Int = 0,
-    @SerialName("pm") val priceModifier: Int = -1,
+    @SerialName("pm") val priceModifier: Long = -1,
 )
