+++ src/main/kotlin/core/model/config/ConfigMasteryChallengeDTO.kt
@@ -5,10 +5,10 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigMasteryChallengeDTO(
-    @SerialName("id") val id: Int = 0,
+    @SerialName("id") val id: Long = 0,
     @SerialName("t") val targetProgress: Int = 0,
     @SerialName("l") val locaID: String = "",
     @SerialName("n") val nextChallenge: ConfigMasteryChallengeDTO? = null,
-    @SerialName("b") val bonusID: Int = 0,
+    @SerialName("b") val bonusID: Long = 0,
     @SerialName("ic") val instantFinishCosts: Int = 0,
 )
