+++ src/main/kotlin/core/model/config/ConfigNeedRequireDTO.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigNeedRequireDTO(
-    @SerialName("nid") val requiredConfigBuildingIds: List<Int> = emptyList(),
+    @SerialName("nid") val requiredConfigBuildingIds: List<Long> = emptyList(),
     @SerialName("a") val configOutputList: List<ConfigOutputDTO> = emptyList(),
 )
