+++ src/main/kotlin/core/model/config/ConfigProfessionalLevelDTO.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigProfessionalLevelDTO(
-    @SerialName("pli") val id: Int = 0,
+    @SerialName("pli") val id: Long = 0,
     @SerialName("xp") val xpThreshold: UInt = 0u,
     @SerialName("sp") val skillpointGain: UInt = 0u,
 )
